Sam/supports readonly (#46)
* make StoryConfig signer optional, create a JsonRpcProvider as fallback

* StoryConfig optional provider

* separate the readonly client and tests into separate files & classes

* use factory functions to create StoryClient instances

* import mocha

* tsconfig

* update generated modules

* add client interfaces

* create read-only resources

* add read-only tests

* Update collectReadOnly.test.ts

* add readonly client tests

* add license read only tests

* tidy imports

* read only ip assets

* update config imports

* relationship read only client & tests

* collect read only test

* update test naming

* remove readonly relationship tests

* license readonly test description

* transaction read-only resource

* update readonly tests with real APIs

* fix missing type

* fix param comment

* fix relationship function order

* merge fix: update type paths

* collect module generated abi file format

* Update README.md

* Update README.md

* Update README.md
