+++ packages/core-sdk/src/resources/collect.ts
@@ -1,40 +1,17 @@
-import { AxiosInstance, AxiosResponse } from "axios";
+import { AxiosInstance } from "axios";
 import { ContractTransaction, Signer } from "ethers";
 
-import {
-  CollectIPAssetRequest,
-  CollectIPAssetResponse,
-  ListCollectionsRequest,
-  ListCollectionsResponse,
-} from "../types/resources/collect";
+import { CollectIPAssetRequest, CollectIPAssetResponse } from "../types/resources/collect";
 import { CollectModule } from "../abi/generated";
 import { handleError } from "../utils/errors";
+import { CollectReadOnlyClient } from "./collectReadOnly";
 
-export class CollectClient {
-  private readonly httpClient: AxiosInstance;
+export class CollectClient extends CollectReadOnlyClient {
   private readonly signer: Signer;
-  private readonly collectModuleContract: CollectModule;
 
   constructor(httpClient: AxiosInstance, signer: Signer, collectModule: CollectModule) {
-    this.httpClient = httpClient;
+    super(httpClient, collectModule);
     this.signer = signer;
-    this.collectModuleContract = collectModule;
-  }
-
-  /**
-   * List all collected IP Assets.
-   *
-   * @returns A Promise that resolves to the ListLicenseResponse.
-   */
-  public async list(request: ListCollectionsRequest): Promise<ListCollectionsResponse> {
-    try {
-      const response: AxiosResponse = await this.httpClient.get(
-        `/collection?franchiseId=${request.franchiseId}`,
-      );
-      return response.data as ListCollectionsResponse;
-    } catch (error: unknown) {
-      handleError(error, `Failed to get collections`);
-    }
   }
 
   /**
