+++ packages/core-sdk/src/resources/ipAsset.ts
@@ -1,30 +1,21 @@
 import { AxiosInstance } from "axios";
 import { Signer } from "ethers";
 
-import {
-  GetIpAssetRequest,
-  GetIpAssetResponse,
-  CreateIpAssetRequest,
-  CreateIpAssetResponse,
-  ListIpAssetRequest,
-  ListIpAssetResponse,
-} from "../types/resources/ipAsset";
+import { CreateIpAssetRequest, CreateIpAssetResponse } from "../types/resources/ipAsset";
 import { handleError } from "../utils/errors";
 import { isIntegerString } from "../utils/utils";
 import { FranchiseRegistry, IpAssetRegistry__factory } from "../abi/generated";
 import { IpAssetRegistry } from "../abi/generated";
+import { IPAssetReadOnlyClient } from "./ipAssetReadOnly";
 
 /**
  * IpAssetClient allows you to create, view, and list IP Assets on Story Protocol.
  */
-export class IPAssetClient {
-  private readonly httpClient: AxiosInstance;
-  private readonly franchiseRegistry: FranchiseRegistry;
+export class IPAssetClient extends IPAssetReadOnlyClient {
   private readonly signer: Signer;
 
   constructor(httpClient: AxiosInstance, franchiseRegistry: FranchiseRegistry, signer: Signer) {
-    this.httpClient = httpClient;
-    this.franchiseRegistry = franchiseRegistry;
+    super(httpClient, franchiseRegistry);
     this.signer = signer;
   }
 
@@ -40,27 +31,6 @@ export class IPAssetClient {
       return ipAssetRegistry;
     } catch (error) {
       handleError(error, "Failed to retrieve IP Asset Registry");
-    }
-  }
-
-  /**
-   * Get an IP Asset based on the specified IP asset ID.
-   *
-   * @param request - the request object for getting an IP Asset.
-   * @returns the response object the contains the fetched IP Asset.
-   */
-  public async get(request: GetIpAssetRequest): Promise<GetIpAssetResponse> {
-    try {
-      if (!isIntegerString(request.ipAssetId)) {
-        throw new Error(`Invalid IP Asset id. Must be an integer. But get: ${request.ipAssetId}`);
-      }
-
-      const response = await this.httpClient.get(
-        `/ipasset/${request.ipAssetId}?franchiseId=${request.franchiseId}`,
-      );
-      return response.data as GetIpAssetResponse;
-    } catch (error: unknown) {
-      handleError(error, "Failed to get IP Asset");
     }
   }
 
@@ -93,20 +63,6 @@ export class IPAssetClient {
       };
     } catch (error) {
       handleError(error, "Failed to create IP Asset");
-    }
-  }
-
-  /**
-   * List all IP assets.
-   *
-   * @returns the response object that contains results from listing query.
-   */
-  public async list(request: ListIpAssetRequest): Promise<ListIpAssetResponse> {
-    try {
-      const response = await this.httpClient.get(`/ipasset?franchiseId=${request.franchiseId}`);
-      return response.data as ListIpAssetResponse;
-    } catch (error) {
-      handleError(error, "Failed to list IP Asset.");
     }
   }
 }
