+++ packages/core-sdk/src/resources/transaction.ts
@@ -1,50 +1,13 @@
-import { AxiosInstance, AxiosResponse } from "axios";
+import { AxiosInstance } from "axios";
 
-import { handleError } from "../utils/errors";
-import {
-  GetTransactionRequest,
-  GetTransactionResponse,
-  ListTransactionResponse,
-} from "../types/resources/transaction";
+import { TransactionReadOnlyClient } from "./transactionReadOnly";
 
 /**
  * TransactionClient allows you to view and monitor transactions on
  * Story Protocol.
  */
-export class TransactionClient {
-  private readonly httpClient: AxiosInstance;
-
+export class TransactionClient extends TransactionReadOnlyClient {
   constructor(httpClient: AxiosInstance) {
-    this.httpClient = httpClient;
-  }
-
-  /**
-   * Get transaction data based on the specified transaction id.
-   *
-   * @param request - the request object for getting the transaction
-   * @returns the response object that contains the fetched transaction object
-   */
-  public async get(request: GetTransactionRequest): Promise<GetTransactionResponse> {
-    try {
-      const response = await this.httpClient.get(`/transaction/${request.txId}`);
-      return response.data as GetTransactionResponse;
-    } catch (error: unknown) {
-      handleError(error, "Failed to get transaction");
-    }
-  }
-
-  /**
-   * Get transaction data based on the specified transaction id.
-   *
-   * @param request - the request object for getting the transactions
-   * @returns the response object that contains the fetched transaction object
-   */
-  public async list(): Promise<ListTransactionResponse> {
-    try {
-      const response: AxiosResponse = await this.httpClient.get(`/transaction`);
-      return response.data as ListTransactionResponse;
-    } catch (error: unknown) {
-      handleError(error, `Failed to get transactions`);
-    }
+    super(httpClient);
   }
 }
