+++ packages/core-sdk/test/integration/license.test.ts
@@ -3,12 +3,13 @@ import chaiAsPromised from "chai-as-promised";
 import { StoryClient, StoryConfig, Environment } from "../../src/index";
 import { ethers } from "ethers";
 import * as dotenv from "dotenv";
+import { Client } from "../../src/types/client";
 
 dotenv.config();
 chai.use(chaiAsPromised);
 
 describe("License Functions", () => {
-  let client: StoryClient;
+  let client: Client;
 
   before(async function () {
     const provider = new ethers.providers.JsonRpcProvider(process.env.RPC_PROVIDER_URL);
@@ -19,44 +20,7 @@ describe("License Functions", () => {
       signer: wallet,
     };
 
-    client = new StoryClient(config);
-  });
-
-  describe("Get License", async function () {
-    it("should return a license when the license id is valid", async () => {
-      await expect(
-        client.license.get({
-          licenseId: "49", // Provide a valid license ID
-        }),
-      ).to.not.be.rejected;
-    });
-
-    it("should handle errors when retrieving a license", async () => {
-      try {
-        await client.license.get({
-          licenseId: "invalid_id", // Provide an invalid license ID
-        });
-        expect.fail("Function should have thrown an error");
-      } catch (error) {}
-    });
-  });
-
-  describe("List Licenses", async function () {
-    const mockListLicenseRequest = {
-      franchiseId: "66",
-      ipAssetId: "49",
-    };
-
-    it("should list all licenses", async () => {
-      await expect(client.license.list(mockListLicenseRequest)).to.not.be.rejected;
-    });
-
-    it("should handle errors when listing licenses", async () => {
-      try {
-        await client.license.list(mockListLicenseRequest);
-        expect.fail("Function should have thrown an error");
-      } catch (error) {}
-    });
+    client = StoryClient.newClient(config);
   });
 
   describe("Create License", async function () {
