+++ packages/core-sdk/test/integration/relationship.test.ts
@@ -3,12 +3,13 @@ import chaiAsPromised from "chai-as-promised";
 import { StoryClient, StoryConfig, Environment } from "../../src/index";
 import { ethers } from "ethers";
 import * as dotenv from "dotenv";
+import { Client } from "../../src/types/client";
 
 dotenv.config();
 chai.use(chaiAsPromised);
 
 describe("Relationship Functions", () => {
-  let client: StoryClient;
+  let client: Client;
   // TODO: need actually IPAsset data for it to work
   const mockRequest = {
     sourceIPAsset: {
@@ -30,7 +31,19 @@ describe("Relationship Functions", () => {
       signer: wallet,
     };
 
-    client = new StoryClient(config);
+    client = StoryClient.newClient(config);
+  });
+
+  describe("Is Relationship Expired", async function () {
+    it("should check if relationship is expired and return result", async () => {
+      await expect(client.relationship.isRelationshipExpired(mockRequest)).to.not.be.rejected;
+    });
+  });
+
+  describe("Is Related", async function () {
+    it("should check if two entities are related and return result", async () => {
+      await expect(client.relationship.isRelated(mockRequest)).to.not.be.rejected;
+    });
   });
 
   describe("Relate", async function () {
@@ -43,18 +56,6 @@ describe("Relationship Functions", () => {
   describe.skip("Unrelate", async function () {
     it("should unrelate and return txHash", async () => {
       await client.relationship.unrelate(mockRequest);
-    });
-  });
-
-  describe("Is Relationship Expired", async function () {
-    it("should check if relationship is expired and return result", async () => {
-      await expect(client.relationship.isRelationshipExpired(mockRequest)).to.not.be.rejected;
-    });
-  });
-
-  describe("Is Related", async function () {
-    it("should check if two entities are related and return result", async () => {
-      await expect(client.relationship.isRelated(mockRequest)).to.not.be.rejected;
     });
   });
 });
